refactor: one print folder for the two document surfaces
src/app/sheet and src/app/capture were two of ten top-level roles for
one world: turning a game renderer into a printable document. AGENTS.md
already described them as a pair while the filesystem kept them apart.

They are now src/app/print/{sheet,capture}. Nine roles instead of ten,
one entry in the layout guard instead of two.

The sheet stays a sibling of capture rather than moving inside it. The
dependency runs one way — delete the publisher and the bridge is still
needed by the preview route and the editor's iframe; delete the bridge
and the publisher cannot render. Two of its three callers capture
nothing, so filing it under capture would have named the folder for the
one caller that does.

"print" rather than "document" because it is the word the repo already
uses for this world — AGENTS.md calls src/game "print-faithful
renderers" and its pieces "print vocabulary" — and because a folder
named document holding files that use the DOM document reads badly.

The two page-root resets drop the now-redundant "document" from their
names: sheet-page.css and capture-page.css. One relative import that
escaped the folder became its @game alias, since it broke on the extra
directory level and would break again on the next move.
