+++ publisher-capture.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/app/capture/publisher-entry.tsx"></script>
+    <script type="module" src="/src/app/print/capture/publisher-entry.tsx"></script>
   </body>
 </html>
