+++ src/app/print/capture/capture-page.css
