+++ src/app/print/capture/publisher-entry.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
-import '@app/sheet/FactionSheetDocument.css';
+import '@app/print/sheet/sheet-page.css';
 import '@app/styles/fonts.css';
 import '@app/styles/tokens.css';
 
@@ -9,7 +9,7 @@ import '@fontsource/caladea/latin-400.css';
 import '@fontsource/caladea/latin-400-italic.css';
 import '@fontsource/caladea/latin-700.css';
 import '@fontsource/caladea/latin-700-italic.css';
-import './capture-document.css';
+import './capture-page.css';
 import { PublisherFactionSheetCapture } from './PublisherFactionSheetCapture';
 
 const root = document.querySelector('#root');
