+++ src/app/print/sheet/FactionSheetView.tsx
