+++ src/app/print/sheet/sheet-page.css
@@ -1,4 +1,7 @@
-/* Neutralize Page.css html decoration when faction sheet route is active (see sheet.tsx). */
+/* Neutralize the app's page decoration while a sheet document is on screen. Keyed on the
+   `data-faction-sheet` attribute, which both document surfaces set on `html`: the preview route
+   (`routes/preview/sheet/$factionSlug.tsx`) and the capture page
+   (`print/capture/PublisherFactionSheetCapture.tsx`). */
 html[data-faction-sheet] {
   background-color: #fff;
   background-image: none;
