+++ src/app/routes/preview/sheet/$factionSlug.tsx
@@ -2,8 +2,8 @@ import { createFileRoute } from '@tanstack/react-router';
 import { useEffect } from 'react';
 
 import { loadFactionBySlug, useFaction } from '@db/factions';
-import '@app/sheet/FactionSheetDocument.css';
-import { FactionSheetView } from '@app/sheet/FactionSheetView';
+import '@app/print/sheet/sheet-page.css';
+import { FactionSheetView } from '@app/print/sheet/FactionSheetView';
 
 import { AssetRenderModeProvider } from '../../../../game/assets/assetRenderMode';
 import { useFactionSheetPostMessage } from './-useFactionSheetPostMessage';
